Moves Models to Models dir
