+++ app/Models/Expense.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use App\Traits\UuidPrimaryKey;
 use Illuminate\Database\Eloquent\Model;
