+++ app/Models/Income.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use App\Traits\UuidPrimaryKey;
 use Illuminate\Database\Eloquent\Model;
 
-class Account extends Model
+class Income extends Model
 {
     use UuidPrimaryKey;
 }
